+++ polyphemus/lib/etls.rb
@@ -9,11 +9,10 @@ require_relative 'etls/sftp_ingest_metis_triage_files_etl'
 require_relative 'etls/sync_cat_files_etl'
 require_relative 'etls/ipi/ipi_create_rna_seq_and_plate_records'
 require_relative 'etls/sftp_ingest_cat_to_c4_triage_files_etl'
-require_relative 'etls/ipi/ipi_rna_seq_link_fastqs_etl'
+require_relative 'etls/ipi/ipi_rna_seq_link_raw_fastq_files_etl'
 require_relative 'etls/ipi/ipi_rna_seq_populate_attributes_etl'
 require_relative 'etls/ipi/ipi_rna_seq_populate_matrices_etl'
-require_relative 'etls/ipi/ipi_rna_seq_link_cram_files_etl'
-require_relative 'etls/ipi/ipi_rna_seq_link_cram_index_files_etl'
-require_relative 'etls/ipi/ipi_rna_seq_link_junction_files_etl'
-require_relative 'etls/ipi/ipi_rna_seq_link_unmapped_fastq_files_etl'
-require_relative 'etls/ipi/ipi_rna_seq_populate_watch_folders_etl'
+require_relative 'etls/ipi/ipi_rna_seq_link_processed_files_etl'
+require_relative 'etls/ipi/ipi_rna_seq_add_processed_files_watch_folders_etl'
+require_relative 'etls/ipi/ipi_rna_seq_add_raw_fastq_files_watch_folders_etl'
+require_relative 'etls/ipi/ipi_rna_seq_add_results_watch_folders_etl'
