+++ polyphemus/lib/polyphemus.rb
@@ -3,7 +3,6 @@ require_relative 'etl_cursor'
 require_relative 'metis_file_etl'
 require_relative 'magma_record_etl'
 require_relative 'etls'
-require_relative 'metis_file_for_magma_model_etl'
 require 'etna/clients/magma'
 require 'etna/clients/metis'
 require 'etna/clients/janus'
