+++ polyphemus/spec/etls/ipi/ipi_rna_seq_add_results_watch_folders_etl_spec.rb
@@ -1,0 +1,108 @@
+describe Polyphemus::IpiRnaSeqAddResultsWatchFoldersEtl do
+  let(:project_name) { "ipi" }
+  let(:bucket_name) { "data" }
+  let(:cursor) {
+    Polyphemus::MetisFolderEtlCursor.new(
+      job_name: "test",
+      project_name: project_name,
+      bucket_name: bucket_name,
+    )
+  }
+  let(:helper) { IpiHelper.new("lib/etls/renaming/projects/test_renames.json") }
+
+  before(:each) do
+    allow(IpiHelper).to receive(:new).and_return(helper)
+    stub_metis_setup
+    copy_renaming_project
+    stub_list_folder(project: project_name, bucket: bucket_name)
+    stub_magma_update_json
+    stub_magma_models(fixture: "spec/fixtures/magma_ipi_models_with_records.json")
+  end
+
+  def folder(folder_name, folder_path, updated_at = Time.now)
+    Etna::Clients::Metis::Folder.new({
+      folder_name: folder_name,
+      folder_path: folder_path,
+      updated_at: updated_at,
+    })
+  end
+
+  describe "create Polyphemus::WatchFile records" do
+    it "for invalid NASH / NAFLD samples" do
+      expect(Polyphemus::WatchFolder.count).to eq(0)
+
+      etl = Polyphemus::IpiRnaSeqAddResultsWatchFoldersEtl.new
+
+      etl.process(cursor, [
+        folder("results", "plate1_rnaseq_new/results"),
+        folder("results", "plate2_rnaseq_new/results"),
+      ])
+
+      expect(Polyphemus::WatchFolder.count).to eq(2)
+    end
+  end
+
+  describe "processes files in found folders" do
+    it "for attributes" do
+      stub_download_file(
+        project: project_name,
+        file_contents: ::File.read("./spec/fixtures/ipi_rna_seq_results.tsv"),
+      )
+      stub_list_folder(
+        project: project_name,
+        bucket: bucket_name,
+        response_body: {
+          files: [{
+            file_name: "rnaseq_table.tsv",
+            download_url: "/#{project_name}/download/rnaseq_table.tsv",
+            bucket_name: bucket_name,
+            project_name: project_name,
+          }],
+          folders: [],
+        }.to_json,
+      )
+
+      etl = Polyphemus::IpiRnaSeqAddResultsWatchFoldersEtl.new
+
+      etl.process(cursor, [
+        folder("results", "plate1_rnaseq_new/results"),
+        folder("results", "plate2_rnaseq_new/results"),
+      ])
+
+      # Make sure rna_seq records are updated. Only when file is found.
+      expect(WebMock).to have_requested(:post, /#{MAGMA_HOST}\/update/)
+    end
+
+    it "for matrices" do
+      require "pry"
+      binding.pry
+      stub_download_file(
+        project: project_name,
+        file_contents: ::File.read("./spec/fixtures/ipi_gene_counts.tsv"),
+      )
+      stub_list_folder(
+        project: project_name,
+        bucket: bucket_name,
+        response_data: {
+          files: [{
+            file_name: "gene_counts_table.tsv",
+            download_url: "/#{project_name}/download/gene_counts_table.tsv",
+            bucket_name: bucket_name,
+            project_name: project_name,
+          }],
+          folders: [],
+        },
+      )
+
+      etl = Polyphemus::IpiRnaSeqAddResultsWatchFoldersEtl.new
+
+      etl.process(cursor, [
+        folder("results", "plate1_rnaseq_new/results"),
+        folder("results", "plate2_rnaseq_new/results"),
+      ])
+
+      # Make sure rna_seq records are updated. Only when file is found.
+      expect(WebMock).to have_requested(:post, /#{MAGMA_HOST}\/update/)
+    end
+  end
+end
